+++ demo/src/main/java/co/edu/udes/activity/backend/demo/repositories/EvaluationRepository.java
@@ -1,14 +1,12 @@
 package co.edu.udes.activity.backend.demo.repositories;
 
 import co.edu.udes.activity.backend.demo.models.Evaluation;
-import co.edu.udes.activity.backend.demo.models.Group;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 import java.util.List;
 
 @Repository
-public interface EvaluationRepository extends JpaRepository<Evaluation, Integer> {
-
+public interface EvaluationRepository extends JpaRepository<Evaluation, Long> {
     List<Evaluation> findByGroupId(long groupId);
 }
